Like sys + validation with annotations (conflicts solved)
